[Refactor] #963 Moved stop_singing() from player-status.h to spells-song.h
