+++ src/mutation/mutation-processor.cpp
@@ -29,6 +29,7 @@
 #include "spell-kind/spells-teleport.h"
 #include "spell-kind/spells-world.h"
 #include "spell-realm/spells-hex.h"
+#include "spell-realm/spells-song.h"
 #include "spell/spell-types.h"
 #include "spell/summon-types.h"
 #include "status/bad-status-setter.h"
