+++ src/player/player-status.h
@@ -48,7 +48,6 @@ bool is_shero(player_type *creature_ptr);
 bool is_echizen(player_type *creature_ptr);
 bool is_in_dungeon(player_type *creature_ptr);
 
-void stop_singing(player_type *creature_ptr);
 void stop_mouth(player_type *caster_ptr);
 PERCENTAGE calculate_upkeep(player_type *creature_ptr);
 bool music_singing(player_type *caster_ptr, int music_songs);
