+++ src/spell-realm/spells-song.h
@@ -5,3 +5,4 @@
 typedef struct player_type player_type;
 void check_music(player_type *caster_ptr);
 bool set_tim_stealth(player_type *creature_ptr, TIME_EFFECT v, bool do_dec);
+void stop_singing(player_type *creature_ptr);
